+++ csrc/services/b2c/msg_read_response_bs.c
@@ -54,7 +54,8 @@ void msg_read_response_bs__alloc_read_response(
     msg_read_resp->NoOfResults = msg_read_response_bs__p_nb_resps;
     if (msg_read_response_bs__p_nb_resps > 0)
     {
-        msg_read_resp->Results = (SOPC_DataValue*) calloc(msg_read_response_bs__p_nb_resps, sizeof(SOPC_DataValue));
+        msg_read_resp->Results =
+            (SOPC_DataValue*) calloc((size_t) msg_read_response_bs__p_nb_resps, sizeof(SOPC_DataValue));
     }
     else
     {
